backend/pyproject.toml
[tool.poetry]
name = "grabber-hr"
version = "0.1.0"
description = ""
authors = ["liam <liam@gravylab.co.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.0"
pyyaml = "^6.0.1"
aiohttp = "^3.8.6"
openai = "^1.35.7"
pytz = "^2023.3.post1"
anthropic = "^0.23.1"
python-mecab-ko = "^1.3.7"
numpy = "^2.1.3"
scikit-learn = "^1.5.2"
selenium = "^4.26.1"
webdriver-manager = "^4.0.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
